refactor: move EnvToUniqueEnv from ddevapp to util package [skip buildkite]
