+++ pkg/util/utils_test.go
@@ -201,6 +201,34 @@ func TestSliceToUniqueSlice(t *testing.T) {
 	}
 }
 
+// TestEnvToUniqueEnv tests EnvToUniqueEnv
+func TestEnvToUniqueEnv(t *testing.T) {
+	assert := asrt.New(t)
+
+	testBedSources := [][]string{
+		{"ONE=one", "ONE=two", "ONE=three", "TWO=two", "TWO=three", "TWO=four"},
+		// Bare variable names (no =value) should pass through for host env lookup by docker-compose
+		{"BARE_VAR", "KEY=value"},
+		// A later KEY=value entry should override an earlier bare KEY entry
+		{"MYVAR", "MYVAR=explicit"},
+		// A later bare KEY entry should override an earlier KEY=value entry
+		{"MYVAR=explicit", "MYVAR"},
+	}
+
+	testBedExpectations := [][]string{
+		{"ONE=three", "TWO=four"},
+		{"BARE_VAR", "KEY=value"},
+		{"MYVAR=explicit"},
+		{"MYVAR"},
+	}
+
+	for i := range testBedSources {
+		res := util.EnvToUniqueEnv(&testBedSources[i])
+		sort.Strings(res)
+		assert.Equal(testBedExpectations[i], res)
+	}
+}
+
 // TestArrayToReadableOutput tests ArrayToReadableOutput
 func TestArrayToReadableOutput(t *testing.T) {
 	assert := asrt.New(t)
